refactor(server): extract routes + bootstrap from server.ts (#4227)
Extracted the server bootstrap (HTTP server setup, service wiring, route
registration, main() function) from a monolithic 690-LOC server.ts into focused
modules, all under the 500-line quality gate.

Files:
- src/server.ts (57 lines) — thin entry point: imports main() and invokes it
- src/server-bootstrap.ts (476 lines) — main() bootstrap + module-level setup
- src/server-http.ts (122 lines) — Fastify app + rate limit + content-type parser
  + request decoration + log sink + security headers
- src/server-inbound.ts (117 lines) — handleInbound() controller (Telegram commands)
- src/server-channels.ts (58 lines) — registerChannels() helper

Acceptance criteria (per Ema's spec):
- No behavior change
- npm run test: 643 files, 8213 tests pass, 26 skipped, 0 failures (was 643/8213/26/0 before)
- npm run build: clean
- npm run gate: clean (all gates pass: arch, hygiene, security, token, audit, lint,
  dashboard:tokens, dashboard:clickable, tsc, build, check-bundle-size, test:serial)
- No edits in extracted files (src/routes/*.ts) beyond imports
- One PR with diff stat showing extraction only

Out of scope: any new route, any new middleware, any refactor of route internals.
